renamed test.py to test_rir.py and remove west related test
